+++ torchvision/models/alexnet.py
@@ -4,12 +4,11 @@ from typing import Any, Optional
 import torch
 import torch.nn as nn
 
+from ..transforms import ImageClassificationEval, InterpolationMode
 from ..utils import _log_api_usage_once
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param
-
-from ..transforms import ImageClassificationEval, InterpolationMode
 
 
 __all__ = ["AlexNet", "AlexNet_Weights", "alexnet"]
