+++ torchvision/models/convnext.py
@@ -7,10 +7,8 @@ from torch.nn import functional as F
 
 from ..ops.misc import Conv2dNormActivation
 from ..ops.stochastic_depth import StochasticDepth
-from ..utils import _log_api_usage_once
-
 from ..transforms import ImageClassificationEval, InterpolationMode
-
+from ..utils import _log_api_usage_once
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param
