+++ torchvision/models/efficientnet.py
@@ -12,7 +12,6 @@ from torchvision.ops import StochasticDepth
 from ..ops.misc import Conv2dNormActivation, SqueezeExcitation
 from ..transforms import ImageClassificationEval, InterpolationMode
 from ..utils import _log_api_usage_once
-
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param, _make_divisible
