+++ torchvision/models/googlenet.py
@@ -1,6 +1,6 @@
 import warnings
-from functools import partial
 from collections import namedtuple
+from functools import partial
 from typing import Optional, Tuple, List, Callable, Any
 
 import torch
@@ -8,9 +8,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch import Tensor
 
-from ..utils import _log_api_usage_once
 from ..transforms import ImageClassificationEval, InterpolationMode
-
+from ..utils import _log_api_usage_once
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param
