+++ torchvision/models/mnasnet.py
@@ -1,5 +1,5 @@
-from functools import partial
 import warnings
+from functools import partial
 from typing import Any, Dict, List, Optional
 
 import torch
