+++ torchvision/models/mobilenetv2.py
@@ -1,5 +1,5 @@
-from functools import partial
 import warnings
+from functools import partial
 from typing import Callable, Any, Optional, List
 
 import torch
@@ -9,7 +9,6 @@ from torch import nn
 from ..ops.misc import Conv2dNormActivation
 from ..transforms import ImageClassificationEval, InterpolationMode
 from ..utils import _log_api_usage_once
-
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param, _make_divisible
