+++ torchvision/models/regnet.py
@@ -9,7 +9,6 @@ from torch import nn, Tensor
 from ..ops.misc import Conv2dNormActivation, SqueezeExcitation
 from ..transforms import ImageClassificationEval, InterpolationMode
 from ..utils import _log_api_usage_once
-
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param, _make_divisible
@@ -781,9 +780,6 @@ class RegNet_X_32GF_Weights(WeightsEnum):
         },
     )
     DEFAULT = IMAGENET1K_V2
-
-
-
 
 
 @handle_legacy_interface(weights=("pretrained", RegNet_Y_400MF_Weights.IMAGENET1K_V1))
