+++ torchvision/models/vgg.py
@@ -6,7 +6,6 @@ import torch.nn as nn
 
 from ..transforms import ImageClassificationEval, InterpolationMode
 from ..utils import _log_api_usage_once
-
 from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param
