+++ torchvision/prototype/models/__init__.py
@@ -1,0 +1,5 @@
+from . import detection
+from . import optical_flow
+from . import quantization
+from . import segmentation
+from . import video
