+++ torchvision/prototype/models/detection/retinanet.py
@@ -1,10 +1,10 @@
 from typing import Any, Optional
 
 from torch import nn
-from torchvision.prototype.transforms import ObjectDetectionEval
-from torchvision.transforms.functional import InterpolationMode
-
-from ....models.detection.retinanet import (
+from torchvision.models._api import WeightsEnum, Weights
+from torchvision.models._meta import _COCO_CATEGORIES
+from torchvision.models._utils import handle_legacy_interface, _ovewrite_value_param
+from torchvision.models.detection.retinanet import (
     _resnet_fpn_extractor,
     _validate_trainable_layers,
     RetinaNet,
@@ -12,10 +12,8 @@ from ....models.detection.retinanet import (
     misc_nn_ops,
     overwrite_eps,
 )
-from torchvision.models._api import WeightsEnum, Weights
-from torchvision.models._meta import _COCO_CATEGORIES
-from .._utils import handle_legacy_interface, _ovewrite_value_param
-from ..resnet import ResNet50_Weights, resnet50
+from torchvision.models.resnet import ResNet50_Weights, resnet50
+from torchvision.transforms import ObjectDetectionEval, InterpolationMode
 
 
 __all__ = [
