+++ torchvision/prototype/models/detection/ssdlite.py
@@ -3,10 +3,10 @@ from functools import partial
 from typing import Any, Callable, Optional
 
 from torch import nn
-from torchvision.prototype.transforms import ObjectDetectionEval
-from torchvision.transforms.functional import InterpolationMode
-
-from ....models.detection.ssdlite import (
+from torchvision.models._api import WeightsEnum, Weights
+from torchvision.models._meta import _COCO_CATEGORIES
+from torchvision.models._utils import handle_legacy_interface, _ovewrite_value_param
+from torchvision.models.detection.ssdlite import (
     _mobilenet_extractor,
     _normal_init,
     _validate_trainable_layers,
@@ -15,10 +15,8 @@ from ....models.detection.ssdlite import (
     SSD,
     SSDLiteHead,
 )
-from torchvision.models._api import WeightsEnum, Weights
-from torchvision.models._meta import _COCO_CATEGORIES
-from .._utils import handle_legacy_interface, _ovewrite_value_param
-from ..mobilenetv3 import MobileNet_V3_Large_Weights, mobilenet_v3_large
+from torchvision.models.mobilenetv3 import MobileNet_V3_Large_Weights, mobilenet_v3_large
+from torchvision.transforms import ObjectDetectionEval, InterpolationMode
 
 
 __all__ = [
