+++ torchvision/prototype/models/quantization/mobilenetv2.py
@@ -1,19 +1,17 @@
 from functools import partial
 from typing import Any, Optional, Union
 
-from torchvision.prototype.transforms import ImageClassificationEval
-from torchvision.transforms.functional import InterpolationMode
-
-from ....models.quantization.mobilenetv2 import (
+from torchvision.models._api import WeightsEnum, Weights
+from torchvision.models._meta import _IMAGENET_CATEGORIES
+from torchvision.models._utils import handle_legacy_interface, _ovewrite_named_param
+from torchvision.models.mobilenetv2 import MobileNet_V2_Weights
+from torchvision.models.quantization.mobilenetv2 import (
     QuantizableInvertedResidual,
     QuantizableMobileNetV2,
     _replace_relu,
     quantize_model,
 )
-from torchvision.models._api import WeightsEnum, Weights
-from torchvision.models._meta import _IMAGENET_CATEGORIES
-from .._utils import handle_legacy_interface, _ovewrite_named_param
-from ..mobilenetv2 import MobileNet_V2_Weights
+from torchvision.transforms import ImageClassificationEval, InterpolationMode
 
 
 __all__ = [
