+++ torchvision/prototype/transforms/_geometry.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, List, Union, Sequence, Tuple, cast
 import PIL.Image
 import torch
 from torchvision.prototype import features
-from torchvision.prototype.transforms import Transform, InterpolationMode, functional as F
-from torchvision.transforms.functional import pil_to_tensor
+from torchvision.prototype.transforms import Transform, functional as F
+from torchvision.transforms.functional import pil_to_tensor, InterpolationMode
 from torchvision.transforms.transforms import _setup_size, _interpolation_modes_from_int
 
 from ._utils import query_image, get_image_dimensions, has_any, is_simple_tensor
