+++ torchvision/prototype/transforms/functional/_geometry.py
@@ -4,9 +4,8 @@ from typing import Tuple, List, Optional, Sequence, Union
 import PIL.Image
 import torch
 from torchvision.prototype import features
-from torchvision.prototype.transforms import InterpolationMode
 from torchvision.transforms import functional_tensor as _FT, functional_pil as _FP
-from torchvision.transforms.functional import pil_modes_mapping, _get_inverse_affine_matrix
+from torchvision.transforms.functional import pil_modes_mapping, _get_inverse_affine_matrix, InterpolationMode
 
 from ._meta import convert_bounding_box_format, get_dimensions_image_tensor, get_dimensions_image_pil
