+++ torchvision/prototype/utils/_internal.py
@@ -1,14 +1,11 @@
 import collections.abc
 import difflib
-import functools
-import inspect
 import io
 import mmap
 import os
 import os.path
 import platform
 import textwrap
-import warnings
 from typing import (
     Any,
     BinaryIO,
@@ -28,31 +25,19 @@ from typing import (
 
 import numpy as np
 import torch
+from torchvision._utils import sequence_to_str
+
 
 __all__ = [
-    "sequence_to_str",
     "add_suggestion",
     "FrozenMapping",
     "make_repr",
     "FrozenBunch",
-    "kwonly_to_pos_or_kw",
     "fromfile",
     "ReadOnlyTensorBuffer",
     "apply_recursively",
     "query_recursively",
 ]
-
-
-def sequence_to_str(seq: Sequence, separate_last: str = "") -> str:
-    if not seq:
-        return ""
-    if len(seq) == 1:
-        return f"'{seq[0]}'"
-
-    head = "'" + "', '".join([str(item) for item in seq[:-1]]) + "'"
-    tail = f"{'' if separate_last and len(seq) == 2 else ','} {separate_last}'{seq[-1]}'"
-
-    return head + tail
 
 
 def add_suggestion(
@@ -149,57 +134,6 @@ class FrozenBunch(FrozenMapping):
 
     def __repr__(self) -> str:
         return make_repr(type(self).__name__, self.items())
-
-
-def kwonly_to_pos_or_kw(fn: Callable[..., D]) -> Callable[..., D]:
-    """Decorates a function that uses keyword only parameters to also allow them being passed as positionals.
-
-    For example, consider the use case of changing the signature of ``old_fn`` into the one from ``new_fn``:
-
-    .. code::
-
-        def old_fn(foo, bar, baz=None):
-            ...
-
-        def new_fn(foo, *, bar, baz=None):
-            ...
-
-    Calling ``old_fn("foo", "bar, "baz")`` was valid, but the same call is no longer valid with ``new_fn``. To keep BC
-    and at the same time warn the user of the deprecation, this decorator can be used:
-
-    .. code::
-
-        @kwonly_to_pos_or_kw
-        def new_fn(foo, *, bar, baz=None):
-            ...
-
-        new_fn("foo", "bar, "baz")
-    """
-    params = inspect.signature(fn).parameters
-
-    try:
-        keyword_only_start_idx = next(
-            idx for idx, param in enumerate(params.values()) if param.kind == param.KEYWORD_ONLY
-        )
-    except StopIteration:
-        raise TypeError(f"Found no keyword-only parameter on function '{fn.__name__}'") from None
-
-    keyword_only_params = tuple(inspect.signature(fn).parameters)[keyword_only_start_idx:]
-
-    @functools.wraps(fn)
-    def wrapper(*args: Any, **kwargs: Any) -> D:
-        args, keyword_only_args = args[:keyword_only_start_idx], args[keyword_only_start_idx:]
-        if keyword_only_args:
-            keyword_only_kwargs = dict(zip(keyword_only_params, keyword_only_args))
-            warnings.warn(
-                f"Using {sequence_to_str(tuple(keyword_only_kwargs.keys()), separate_last='and ')} as positional "
-                f"parameter(s) is deprecated. Please use keyword parameter(s) instead."
-            )
-            kwargs.update(keyword_only_kwargs)
-
-        return fn(*args, **kwargs)
-
-    return wrapper
 
 
 def _read_mutable_buffer_fallback(file: BinaryIO, count: int, item_size: int) -> bytearray:
